make tracing enabled a build time flag
