+++ bbq/vm/context.go
@@ -383,7 +383,3 @@ func (c *Context) MaybeUpdateStorageReferenceMemberReceiver(
 
 	return member
 }
-
-func (c *Context) TracingEnabled() bool {
-	return c.Tracer != nil
-}
