+++ interpreter/interpreter_tracing_test.go
@@ -1,3 +1,5 @@
+//go:build cadence_tracing
+
 /*
  * Cadence - The resource-oriented smart contract programming language
  *
@@ -72,9 +74,8 @@ func prepareWithTracingCallBack(
 			nil,
 			TestLocation,
 			&interpreter.Config{
-				Storage:        storage,
-				TracingEnabled: true,
-				OnRecordTrace:  onRecordTrace,
+				Storage:       storage,
+				OnRecordTrace: onRecordTrace,
 			},
 		)
 		require.NoError(t, err)
