+++ interpreter/tracing_disabled.go
@@ -1,3 +1,5 @@
+//go:build !cadence_tracing
+
 /*
  * Cadence - The resource-oriented smart contract programming language
  *
@@ -16,21 +18,6 @@
  * limitations under the License.
  */
 
-package runtime
+package interpreter
 
-import (
-	"github.com/onflow/cadence/interpreter"
-)
-
-// Config is a constant/read-only configuration of an environment.
-type Config struct {
-	Debugger *interpreter.Debugger
-	// StackDepthLimit specifies the maximum depth for call stacks
-	StackDepthLimit uint64
-	// AtreeValidationEnabled configures if atree validation is enabled
-	AtreeValidationEnabled bool
-	// ResourceOwnerChangeCallbackEnabled configures if the resource owner change callback is enabled
-	ResourceOwnerChangeHandlerEnabled bool
-	// CoverageReport enables and collects coverage reporting metrics
-	CoverageReport *CoverageReport
-}
+const TracingEnabled = false
