+++ interpreter/value_array.go
@@ -98,7 +98,7 @@ func NewArrayValueWithIterator(
 
 	var v *ArrayValue
 
-	if context.TracingEnabled() {
+	if TracingEnabled {
 		startTime := time.Now()
 
 		defer func() {
@@ -359,7 +359,7 @@ func (v *ArrayValue) Destroy(context ResourceDestructionContext, locationRange L
 		},
 	)
 
-	if context.TracingEnabled() {
+	if TracingEnabled {
 		startTime := time.Now()
 
 		valueID := v.ValueID().String()
@@ -1198,7 +1198,7 @@ func (v *ArrayValue) ConformsToStaticType(
 	results TypeConformanceResults,
 ) bool {
 
-	if context.TracingEnabled() {
+	if TracingEnabled {
 		startTime := time.Now()
 
 		valueID := v.ValueID().String()
@@ -1329,7 +1329,7 @@ func (v *ArrayValue) Transfer(
 		},
 	)
 
-	if context.TracingEnabled() {
+	if TracingEnabled {
 		startTime := time.Now()
 
 		valueID := v.ValueID().String()
@@ -1501,7 +1501,7 @@ func (v *ArrayValue) Clone(context ValueCloneContext) Value {
 }
 
 func (v *ArrayValue) DeepRemove(context ValueRemoveContext, hasNoParentContainer bool) {
-	if context.TracingEnabled() {
+	if TracingEnabled {
 		startTime := time.Now()
 
 		valueID := v.ValueID().String()
